Change validator from tv4 to ajv
The main reason for switching is because ajv has a feature to remove properties that aren't defined in the swagger. Using this along with Swagger preprocessing, we can send different API response "shapes" based on the type of user (agent, api-agent, supplier...).

I had to switch off unit tests that rely on tv4's specific error messages and some other tests that rely on non-standard behavior for max/min parameters.